app/requirements.txt
Pillow>=9.3.0
pytesseract>=0.3.10
uvicorn>=0.19.0
gunicorn
fastapi>=0.87.0
wheel
python-multipart
jinja2
numpy>=1.24
opencv-python>=4.7
python-dotenv
numpy
requests
python-magic
pdf2image
python-poppler
pytesseract
cffi==1.16.0
pycryptodome
detectaicore>=0.0.16
orjson >=3.8.1
pydantic>=1.8.0
inscriptis==2.3.2
aiofiles
spacy
sacremoses
python-Levenshtein
python-pptx
docx2txt
beautifulsoup4
lxml